Sign Out and Adding Social
Implemented signing out. Right now, the token isn't invalidated, but it is removed from the client (this is temporary).

Also had to fix a bug with adding social that removed the social.id from the database causing the user to not be able to log into his/her profile again