Merge accounts. Bug reporting
Users can now merge accounts. If the systems detects that a user is trying to add a social to an account and that social is already associated with another account, users can now merge the two accounts together.

Added a link to the discourses so that beta users can report bugs